Deleting works successful

+++ src/components/DeleteButton.js
@@ -1,21 +1,53 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Button } from '@chakra-ui/react';
-import {TfiTrash } from "react-icons/tfi";
+import { TfiTrash } from "react-icons/tfi";
 
-export function DeleteButton ({workId, onDelete}) {
-    const handleDeleteClick = () => {
-        fetch(`http://localhost:9292/works/${workId}`, {
-            method: 'DELETE'
-        })
-        .then(()=>{
-            onDelete(workId);
-        })
-        .catch(error => {
-            console.error('Error:', error);
-        });
-    }
+export function DeleteButton(id) {
+    const [workId, setWorkId] = useState(id)
 
-    <Button leftIcon={<TfiTrash />} colorScheme='green' variant='solid' onClick={handleDeleteClick}>
-        Delete
-    </Button>
+    useEffect(() => {
+        const handleDeleteWork = async () => {
+          if (workId) {
+            try {
+              const response = await fetch(`http://localhost:9292/works/${workId}`, {
+                method: "DELETE"
+              });
+    
+              if (response.ok) {
+                console.log("Work deleted successfully");
+                setWorkId('');
+              } else {
+                console.log("Failed to delete work");
+              }
+            } catch (error) {
+              console.error("Error deleting work:", error);
+            }
+          }
+        };
+        handleDeleteWork();
+      }, [workId]);
+
+    return (
+        <Button leftIcon={<TfiTrash />} colorScheme='green' variant='solid' disabled={!workId}>
+          Delete
+        </Button>
+    )
 }
+
+//export function DeleteButton ({workId, onDelete}) {
+  //  const handleDeleteClick = () => {
+    //    fetch(`http://localhost:9292/works/${workId}`, {
+      //      method: 'DELETE'
+        //})
+        //.then(()=>{
+          //  onDelete(workId);
+        //})
+        //.catch(error => {
+          //  console.error('Error:', error);
+        //});
+    //}
+
+    //<Button leftIcon={<TfiTrash />} colorScheme='green' variant='solid' onClick={handleDeleteClick}>
+      //  Delete
+    //</Button>
+//}
